Fix typo
While I'm at it, the "Archives for all past versions of Solr are available at <a href="https://archive.apache.org/dist/solr/">the Apache archives</a>" part isn't correct. That URL gives a 404, and I assume only 9.0+ versions would go there? Or all of them?

Or can I ignore the 404 for now, as this is WIP?